+++ Changelog.md
@@ -1,6 +1,7 @@
 ## 0.3.1 (unreleased)
 
 - Fix UTF-8 decoding of incomplete UTF-8 multibyte sequences to properly report `Invalid`.
+- Change signature of `inline from_utf8(string_view const&)` slightly by dropping its cref.
 
 ## 0.3.0 (2023-03-01)